VPP apps deployed for same country code as Intune
VPP apps are displayed/deployed only for the same country code as your Intune account. Intune will only sync apps (and thus be able to deploy) from the same iTunes locale as the Intune account country code. In other words, if you purchase a particular app which is only available in the US store for iTunes, but your Intune account is German, and the app is not available in Germany, Intune will not show that app (because the app is not available in Germany).