Update windows-update-for-business-configure.md
In my opinion the change back to ProviderID is not correct. We had a SR open for it and the PG told us to use MS DM Server. After the change it worked without problem. This is about 6 month ago. I still have MS DM Server in all implementations and it's working. I suggest please review again, maybe with the PG for Intune and Windows 10.