Updating info on device check in notifications
Doc used to only say that devices receive notifications to check in upon new assignment of policies or apps. 

Actual behavior is that a device is notified whenever there is *any change* in Policy(Create/Update/Delete), change in assignment, or change in the groups(user/device moving in and out of a group).